+++ src/pluginManager.test.ts
@@ -66,7 +66,7 @@ describe('PluginManager', () => {
   let plugins: PluginManager;
   let devices: DeviceManager;
 
-  let parseSpy: jest.SpiedFunction<typeof JSON.parse>;
+  let jsonParseSpy: jest.SpiedFunction<typeof JSON.parse>;
 
   beforeAll(async () => {
     //
@@ -74,14 +74,14 @@ describe('PluginManager', () => {
 
   beforeEach(async () => {
     // Spy on JSON.parse
-    parseSpy = jest.spyOn(JSON, 'parse');
+    jsonParseSpy = jest.spyOn(JSON, 'parse');
     // Clear all mocks
     jest.clearAllMocks();
   });
 
   afterEach(async () => {
     // Restore JSON.parse
-    parseSpy.mockRestore();
+    jsonParseSpy.mockRestore();
   });
 
   afterAll(async () => {
@@ -96,14 +96,14 @@ describe('PluginManager', () => {
     expect(plugins).toBeInstanceOf(PluginManager);
     devices = matterbridge.devices;
     expect(devices).toBeInstanceOf(DeviceManager);
-  }, 60000);
+  });
 
   test('logLevel changes correctly', () => {
     plugins.logLevel = LogLevel.DEBUG;
     expect((plugins as any).log.logLevel).toBe(LogLevel.DEBUG);
   });
 
-  test('clear and load from storage', async () => {
+  test('clear, save and load from storage', async () => {
     plugins.clear();
     expect(await plugins.saveToStorage()).toBe(0);
     expect(await plugins.loadFromStorage()).toHaveLength(0);
@@ -116,6 +116,14 @@ describe('PluginManager', () => {
     await expect(plugins.saveToStorage()).rejects.toThrow(new Error('loadFromStorage: node context is not available.'));
     await expect(plugins.loadFromStorage()).rejects.toThrow(new Error('loadFromStorage: node context is not available.'));
     matterbridge.nodeContext = context;
+  });
+
+  test('async forEach to return', async () => {
+    let count = 0;
+    await plugins.forEach(async (plugin: RegisteredPlugin) => {
+      count++;
+    });
+    expect(count).toBe(0);
   });
 
   test('size returns correct number of plugins', () => {
@@ -132,6 +140,11 @@ describe('PluginManager', () => {
       { 'author': 'To update', 'description': 'To update', homepage: 'https://example.com', 'name': 'matterbridge-mock2', 'path': './src/mock/plugin2/package.json', 'type': 'Unknown', 'version': '1.0.0' },
       { 'author': 'To update', 'description': 'To update', homepage: 'https://example.com', 'name': 'matterbridge-mock3', 'path': './src/mock/plugin3/package.json', 'type': 'Unknown', 'version': '1.0.0' },
     ]);
+  });
+
+  test('save and load from storage', async () => {
+    expect(await plugins.saveToStorage()).toBe(3);
+    expect(await plugins.loadFromStorage()).toHaveLength(3);
   });
 
   test('has returns true if plugin exists', () => {
@@ -608,34 +621,36 @@ describe('PluginManager', () => {
     expect(plugins.length).toBe(0);
   });
 
-  test('add/disable/enable/remove plugin matterbridge-eve-door', async () => {
-    execSync('npm install -g matterbridge-eve-door --omit=dev');
-
+  test('install example plugins', async () => {
+    execSync('npm install -g matterbridge-example-accessory-platform --omit=dev');
+    execSync('npm install -g matterbridge-example-dynamic-platform --omit=dev');
     expect(plugins.length).toBe(0);
-    expect(await plugins.add('matterbridge-eve-door')).not.toBeNull();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Added plugin ${plg}matterbridge-eve-door${nf}`);
+  }, 60000);
+
+  test('add/disable/enable/remove plugin matterbridge-example-accessory-platform', async () => {
+    expect(plugins.length).toBe(0);
+    expect(await plugins.add('matterbridge-example-accessory-platform')).not.toBeNull();
+    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Added plugin ${plg}matterbridge-example-accessory-platform${nf}`);
     expect(plugins.length).toBe(1);
-    const plugin = plugins.get('matterbridge-eve-door');
+    const plugin = plugins.get('matterbridge-example-accessory-platform');
     expect(plugin).not.toBeUndefined();
-    expect(plugin?.name).toBe('matterbridge-eve-door');
+    expect(plugin?.name).toBe('matterbridge-example-accessory-platform');
     expect(plugin?.enabled).toBe(true);
 
-    expect(await plugins.disable('matterbridge-eve-door')).not.toBeNull();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Disabled plugin ${plg}matterbridge-eve-door${nf}`);
+    expect(await plugins.disable('matterbridge-example-accessory-platform')).not.toBeNull();
+    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Disabled plugin ${plg}matterbridge-example-accessory-platform${nf}`);
     expect(plugins.length).toBe(1);
     expect(plugin?.enabled).toBe(false);
 
-    expect(await plugins.enable('matterbridge-eve-door')).not.toBeNull();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Enabled plugin ${plg}matterbridge-eve-door${nf}`);
+    expect(await plugins.enable('matterbridge-example-accessory-platform')).not.toBeNull();
+    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Enabled plugin ${plg}matterbridge-example-accessory-platform${nf}`);
     expect(plugins.length).toBe(1);
     expect(plugin?.enabled).toBe(true);
 
-    expect(await plugins.remove('matterbridge-eve-door')).not.toBeNull();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Removed plugin ${plg}matterbridge-eve-door${nf}`);
+    expect(await plugins.remove('matterbridge-example-accessory-platform')).not.toBeNull();
+    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Removed plugin ${plg}matterbridge-example-accessory-platform${nf}`);
     expect(plugins.length).toBe(0);
-
-    execSync('npm uninstall -g matterbridge-eve-door');
-  }, 300000); // Fails randomly on windows workflows that are really slow
+  });
 
   test('save to storage', async () => {
     (plugins as any)._plugins.set('matterbridge-mock1', { name: 'matterbridge-mock1', path: './src/mock/plugin1/package.json', type: 'Unknown', version: '1.0.0', description: 'To update', author: 'To update' });
@@ -646,33 +661,13 @@ describe('PluginManager', () => {
     expect(await plugins.saveToStorage()).toBe(0);
   });
 
-  test('install plugin matterbridge-xyz', async () => {
-    expect(await plugins.install('matterbridge-xyz')).toBeUndefined();
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Installing plugin ${plg}matterbridge-xyz${nf}`);
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining(`Failed to install plugin ${plg}matterbridge-xyz${er}:`));
-  }, 300000);
-
-  test('install plugin matterbridge-example-accessory-platform', async () => {
-    const version = await plugins.install('matterbridge-example-accessory-platform');
-    expect(version).not.toBeUndefined();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Installing plugin ${plg}matterbridge-example-accessory-platform${nf}`);
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Installed plugin ${plg}matterbridge-example-accessory-platform${nf}`);
-  }, 300000);
-
-  test('install plugin matterbridge-example-dynamic-platform', async () => {
-    const version = await plugins.install('matterbridge-example-dynamic-platform');
-    expect(version).not.toBeUndefined();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Installing plugin ${plg}matterbridge-example-dynamic-platform${nf}`);
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Installed plugin ${plg}matterbridge-example-dynamic-platform${nf}`);
-  }, 300000);
-
   test('add plugin matterbridge-xyz', async () => {
     expect(plugins.length).toBe(0);
     const plugin = await plugins.add('matterbridge-xyz');
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, `Failed to add plugin ${plg}matterbridge-xyz${er}: package.json not found`);
     expect(plugin).toBeNull();
     expect(plugins.length).toBe(0);
-  }, 60000);
+  });
 
   test('add plugin matterbridge-example-accessory-platform', async () => {
     expect(plugins.length).toBe(0);
@@ -695,7 +690,7 @@ describe('PluginManager', () => {
     if (!newPlugin) return;
     expect(newPlugin.name).toBe('matterbridge-example-accessory-platform');
     expect(newPlugin.type).toBe('AnyPlatform');
-  }, 60000);
+  });
 
   test('add plugin matterbridge-example-dynamic-platform', async () => {
     expect(plugins.length).toBe(1);
@@ -709,7 +704,7 @@ describe('PluginManager', () => {
     if (!newPlugin) return;
     expect(newPlugin.name).toBe('matterbridge-example-dynamic-platform');
     expect(newPlugin.type).toBe('AnyPlatform');
-  }, 60000);
+  });
 
   test('load default config plugin matterbridge-example-accessory-platform', async () => {
     const configFileName = path.join(matterbridge.matterbridgeDirectory, `matterbridge-example-accessory-platform.config.json`);
@@ -756,7 +751,7 @@ describe('PluginManager', () => {
     expect(config.type).toBe('AnyPlatform');
     expect(config.debug).toBeDefined();
     expect(config.unregisterOnShutdown).toBeDefined();
-  }, 60000);
+  });
 
   test('save config from plugin matterbridge-example-accessory-platform', async () => {
     expect(plugins.length).toBe(2);
@@ -789,7 +784,7 @@ describe('PluginManager', () => {
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Saved config file ${configFile} for plugin ${plg}${plugin.name}${db}`);
 
     await plugins.shutdown(plugin, 'Test with Jest', true, true);
-  }, 60000);
+  });
 
   test('save config from json matterbridge-example-accessory-platform', async () => {
     expect(plugins.length).toBe(2);
@@ -822,7 +817,7 @@ describe('PluginManager', () => {
     expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.DEBUG, `Saved config file ${configFile} for plugin ${plg}${plugin.name}${db}`);
 
     await plugins.shutdown(plugin, 'Test with Jest', true, true);
-  }, 60000);
+  });
 
   test('load schema plugin matterbridge-example-accessory-platform', async () => {
     expect(plugins.length).toBe(2);
@@ -834,8 +829,7 @@ describe('PluginManager', () => {
       await fs.writeFile(schemaFile, JSON.stringify(plugins.getDefaultSchema(plugin), null, 2), 'utf8');
     } catch (error) {
       // Ignore error
-      // eslint-disable-next-line no-console
-      console.error('writeFile:', schemaFile, error);
+      // console.error('writeFile:', schemaFile, error);
     }
 
     const schema = await plugins.loadSchema(plugin);
@@ -850,7 +844,7 @@ describe('PluginManager', () => {
     expect((schema.properties as any).type).toBeDefined();
     expect((schema.properties as any).debug).toBeDefined();
     expect((schema.properties as any).unregisterOnShutdown).toBeDefined();
-  }, 60000);
+  });
 
   test('load default schema plugin matterbridge-example-accessory-platform', async () => {
     expect(plugins.length).toBe(2);
@@ -862,8 +856,7 @@ describe('PluginManager', () => {
       await fs.unlink(schemaFile);
     } catch (error) {
       // Ignore error
-      // eslint-disable-next-line no-console
-      console.error('unlink:', schemaFile, error);
+      // console.error('unlink:', schemaFile, error);
     }
 
     const schema = await plugins.loadSchema(plugin);
@@ -878,7 +871,7 @@ describe('PluginManager', () => {
     expect((schema.properties as any).type).toBeDefined();
     expect((schema.properties as any).debug).toBeDefined();
     expect((schema.properties as any).unregisterOnShutdown).toBeDefined();
-  }, 60000);
+  });
 
   test('should not load plugin matterbridge-example-accessory-platform if parse error', async () => {
     const plugin = plugins.get('matterbridge-example-accessory-platform');
@@ -899,7 +892,7 @@ describe('PluginManager', () => {
     plugin.platform = savedPlatform;
 
     loggerLogSpy.mockClear();
-    parseSpy.mockImplementationOnce(() => {
+    jsonParseSpy.mockImplementationOnce(() => {
       throw new Error('Test error');
     });
     const platform = await plugins.load(plugin, false, 'Test with Jest');
@@ -1032,14 +1025,6 @@ describe('PluginManager', () => {
 
     plugin = await plugins.start(plugin, 'Test with Jest', false);
 
-    // Log all calls to loggerLogSpy
-    /*
-    consoleLogSpy.mockRestore();
-    loggerLogSpy.mock.calls.forEach((call, index) => {
-      console.log(`Call ${index + 1}:`, call);
-    });
-    */
-
     expect(plugin).toBeDefined();
     expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Starting plugin ${plg}${plugin?.name}${nf} type ${typ}${plugin?.type}${nf}`);
     expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.NOTICE, `Started plugin ${plg}${plugin?.name}${nt} type ${typ}${plugin?.type}${nt}`);
@@ -1055,7 +1040,7 @@ describe('PluginManager', () => {
       true,
     );
     expect(plugin.started).toBe(true);
-  }, 60000);
+  });
 
   test('configure plugin matterbridge-example-accessory-platform should log errors', async () => {
     const plugin = plugins.get('matterbridge-example-accessory-platform');
@@ -1085,7 +1070,7 @@ describe('PluginManager', () => {
     expect(await plugins.configure(plugin)).toBeUndefined();
     plugin.configured = false;
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Plugin ${plg}${plugin?.name}${db} already configured`);
-  }, 60000);
+  });
 
   test('configure plugin matterbridge-example-accessory-platform should log if it throws', async () => {
     const plugin = plugins.get('matterbridge-example-accessory-platform');
@@ -1103,7 +1088,7 @@ describe('PluginManager', () => {
     }
     await plugins.configure(plugin);
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining(`Failed to configure plugin ${plg}${plugin.name}${er}: Test error`));
-  }, 60000);
+  });
 
   test('configure plugin matterbridge-example-accessory-platform', async () => {
     let plugin = plugins.get('matterbridge-example-accessory-platform');
@@ -1119,7 +1104,7 @@ describe('PluginManager', () => {
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.NOTICE, `Configured plugin ${plg}${plugin?.name}${nt} type ${typ}${plugin?.type}${nt}`);
     expect(plugin).not.toBeUndefined();
     expect(plugin.configured).toBe(true);
-  }, 60000);
+  });
 
   test('should not shutdown plugin matterbridge-example-accessory-platform', async () => {
     const plugin = plugins.get('matterbridge-example-accessory-platform');
@@ -1197,7 +1182,7 @@ describe('PluginManager', () => {
     expect(plugin.platform).toBe(undefined);
 
     expect(await plugins.saveToStorage()).toBe(2);
-  }, 60000);
+  });
 
   test('load, start and configure in parallel plugin matterbridge-example-dynamic-platform', async () => {
     const plugin = plugins.get('matterbridge-example-dynamic-platform');
@@ -1208,8 +1193,6 @@ describe('PluginManager', () => {
     config.whiteList = ['No devices'];
     await plugins.saveConfigFromJson(plugin, config);
 
-    // loggerLogSpy.mockRestore();
-    // consoleLogSpy.mockRestore();
     plugins.load(plugin, true, 'Test with Jest', true);
     expect(plugin.loaded).toBe(undefined);
     expect(plugin.started).toBe(undefined);
@@ -1239,57 +1222,13 @@ describe('PluginManager', () => {
 
     plugin = await plugins.shutdown(plugin, 'Test with Jest', true);
     expect(plugin).not.toBeUndefined();
+  });
+
+  test('uninstall example plugins', async () => {
+    execSync('npm uninstall -g matterbridge-example-accessory-platform --omit=dev');
+    execSync('npm uninstall -g matterbridge-example-dynamic-platform --omit=dev');
+    expect(plugins.length).toBe(2);
   }, 60000);
-
-  test('uninstall not existing plugin matterbridge-xyz', async () => {
-    expect(await plugins.uninstall('matterbridge-xyz')).toBeDefined();
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Uninstalling plugin ${plg}matterbridge-xyz${nf}`);
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Uninstalled plugin ${plg}matterbridge-xyz${nf}`);
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, expect.stringContaining(`Uninstalled plugin ${plg}matterbridge-xyz${db}:`));
-  }, 300000);
-
-  test('uninstall plugin matterbridge-example-accessory-platform', async () => {
-    expect(await plugins.uninstall('matterbridge-example-accessory-platform')).toBeDefined();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Uninstalling plugin ${plg}matterbridge-example-accessory-platform${nf}`);
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Uninstalled plugin ${plg}matterbridge-example-accessory-platform${nf}`);
-  }, 300000);
-
-  test('uninstall plugin matterbridge-example-dynamic-platform', async () => {
-    expect(await plugins.uninstall('matterbridge-example-dynamic-platform')).toBeDefined();
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Uninstalling plugin ${plg}matterbridge-example-dynamic-platform${nf}`);
-    expect((plugins as any).log.log).toHaveBeenCalledWith(LogLevel.INFO, `Uninstalled plugin ${plg}matterbridge-example-dynamic-platform${nf}`);
-  }, 300000);
-
-  test('install not existing plugin matterbridge-xyz with mock', async () => {
-    expect(await plugins.install('matterbridge-xyz')).toBeUndefined();
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining(`Failed to install plugin ${plg}matterbridge-xyz${er}:`));
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, expect.stringContaining(`Failed to install plugin ${plg}matterbridge-xyz${db}:`));
-
-    (exec as any).mockImplementation((command: string, callback?: (error: ExecException | null, stdout: string, stderr: string) => void) => {
-      if (command.includes('npm list -g matterbridge-xyz --depth=0')) {
-        if (callback) callback(new Error('Test error'), '', '');
-      } else {
-        if (callback) callback(null, '', '');
-      }
-      // return Promise.resolve() as unknown as ChildProcess;
-    });
-    expect(await plugins.install('matterbridge-xyz')).toBeUndefined();
-  }, 300000);
-
-  test('uninstall not existing plugin matterbridge-xyz with mock', async () => {
-    expect(await plugins.uninstall('matterbridge-xyz')).toBeDefined();
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, expect.stringContaining(`Uninstalled plugin ${plg}matterbridge-xyz${nf}`));
-
-    (exec as any).mockImplementation((command: string, callback?: (error: ExecException | null, stdout: string, stderr: string) => void) => {
-      if (command.includes('npm uninstall -g matterbridge-xyz')) {
-        if (callback) callback(new Error('Test error'), '', '');
-      } else {
-        if (callback) callback(null, '', '');
-      }
-      // return Promise.resolve() as unknown as ChildProcess;
-    });
-    expect(await plugins.uninstall('matterbridge-xyz')).toBeUndefined();
-  }, 300000);
 
   test('Matterbridge.destroyInstance()', async () => {
     // Close the Matterbridge instance